Use system calls to manage Windows services
