+++ pkg/util/initsystem/BUILD
@@ -7,12 +7,19 @@ load(
 
 go_library(
     name = "go_default_library",
-    srcs = ["initsystem.go"],
-    importpath = "k8s.io/kubernetes/pkg/util/initsystem",
-    deps = [
-        "//vendor/golang.org/x/sys/windows/svc:go_default_library",
-        "//vendor/golang.org/x/sys/windows/svc/mgr:go_default_library",
+    srcs = [
+        "initsystem.go",
+        "initsystem_unix.go",
+        "initsystem_windows.go",
     ],
+    importpath = "k8s.io/kubernetes/pkg/util/initsystem",
+    deps = select({
+        "@io_bazel_rules_go//go/platform:windows": [
+            "//vendor/golang.org/x/sys/windows/svc:go_default_library",
+            "//vendor/golang.org/x/sys/windows/svc/mgr:go_default_library",
+        ],
+        "//conditions:default": [],
+    }),
 )
 
 filegroup(
